tratamento de exceções personalizadas

+++ backend/src/main/java/com/dev/sistema/relatorio/service/impl/ReportServiceImpl.java
@@ -6,6 +6,9 @@ import java.util.UUID;
 import java.util.logging.Logger;
 
 import com.dev.sistema.relatorio.dto.ReportDTO;
+import com.dev.sistema.relatorio.exception.ReportFormInvalidFieldsException;
+import com.dev.sistema.relatorio.exception.ReportNoContentException;
+import com.dev.sistema.relatorio.exception.ReportNotFoundException;
 import com.dev.sistema.relatorio.mapper.AttachmentMapper;
 import com.dev.sistema.relatorio.mapper.ReportMapper;
 import com.dev.sistema.relatorio.model.Attachment;
@@ -38,39 +41,43 @@ class ReportServiceImpl implements ReportService {
         try {
             savedReport = repository.save(reportEntity);
 
-            List<Attachment> attachmentList = AttachmentMapper.toEntityList(reportDto.getAttachments());
-            attachmentList.forEach(attachment -> {
-                attachment.setReport(savedReport);
-                attachmentService.saveAttachment(attachment);
-            });
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to save report", e);
+            reportDto.getAttachments()
+                    .stream()
+                    .map(AttachmentMapper::toEntity)
+                    .filter(attachment -> attachment.getId() == null)
+                    .forEach(attachment -> {
+                        attachment.setReport(savedReport);
+                        attachmentService.saveAttachment(attachment);
+                    });
+        } catch (ReportFormInvalidFieldsException exception) {
+            throw new ReportFormInvalidFieldsException();
         }
 
         return savedReport;
     }
 
     @Override
-    public Page<Report> getAllReports(String search, Integer page, Integer pageSize) {
+    public Page<ReportDTO> getAllReports(String search, Integer page, Integer pageSize) {
         Pageable pageable = PageRequest.of(page, pageSize);
         Page<Report> reports;
         try {
-            if(search != null){
+            if (search != null) {
                 reports = repository.findAllBySearch(search, pageable);
             } else {
                 reports = repository.findAll(pageable);
             }
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+        } catch (ReportNoContentException exception) {
+            throw new ReportNoContentException("Failed to list reports.");
         }
-        return reports;
+
+        return reports.map(ReportMapper::toDto);
     }
 
     @Override
     public ReportDTO getReportById(UUID id) {
         Optional<Report> reportSearched = repository.findById(id);
         if (reportSearched.isEmpty()) {
-            throw new RuntimeException("Failed to find report.");
+            throw new ReportNotFoundException("Failed to find report with id " + id);
         }
         return ReportMapper.toDto(reportSearched.get());
     }
@@ -78,40 +85,33 @@ class ReportServiceImpl implements ReportService {
     @Override
     @Transactional
     public void updateReport(ReportDTO reportDto, UUID id) {
-        Optional<Report> reportSearched = repository.findById(id);
-        if (reportSearched.isEmpty()) {
-            throw new RuntimeException("Failed to find report.");
-        }
+        Report reportSearched = repository.findById(id).orElseThrow(() -> new ReportNotFoundException("Failed to find report."));
 
         Report updatedReport = ReportMapper.toEntity(reportDto);
-        updatedReport.setId(reportSearched.get().getId());
-        updatedReport.setCreatedAt(reportSearched.get().getCreatedAt());
+        updatedReport.setId(reportSearched.getId());
+        updatedReport.setCreatedAt(reportSearched.getCreatedAt());
         try {
             repository.save(updatedReport);
 
-            List<Attachment> attachmentList = AttachmentMapper.toEntityList(reportDto.getAttachments());
-            attachmentList.forEach(attachment -> {
-                if (attachment.getId() == null) {
-                    attachment.setReport(updatedReport);
-                    attachmentService.saveAttachment(attachment);
-                }
-            });
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to update report.");
+            reportDto.getAttachments()
+                    .stream()
+                    .map(AttachmentMapper::toEntity)
+                    .filter(attachment -> attachment.getId() == null)
+                    .forEach(attachment -> {
+                        attachment.setReport(updatedReport);
+                        attachmentService.saveAttachment(attachment);
+                    });
+        } catch (ReportNotFoundException e) {
+            throw new ReportNotFoundException("Failed to update report.");
         }
     }
 
     @Override
     @Transactional
     public void deleteReportById(UUID id) {
-        Optional<Report> reportSearched = repository.findById(id);
-        if (reportSearched.isEmpty()) {
-            Logger.getLogger("Failed to find report");
+        if (!repository.existsById(id)) {
+            throw new ReportNotFoundException("Failed to find report.");
         }
-        try {
-            repository.deleteById(id);
-        } catch (Exception e) {
-            reportSearched.ifPresent(report -> Logger.getLogger("Failed to delete report."));
-        }
+        repository.deleteById(id);
     }
 }
